% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usu_data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Team Name Conversion Dictionary (from ncaahoopR)}
\format{
A data frame with 9 columns:
\describe{
\item{NCAA}{Team name as listed on the NCAA website}
\item{ESPN}{Team name as listed in ESPN URLs}
\item{ESPN_PBP}{Team name as listed on ESPN play-by-play logs}
\item{Warren_Nolan}{Team name from WarrenNolan.com}
\item{Trank}{Team name from barttorvik.com}
\item{name_247}{Team name from 247Sports.com}
\item{conference}{The team's conference}
\item{sref_link}{Slug for the team on sports-reference.com}
\item{sref_name}{Team name on sports-reference.com}
}
}
\source{
Extracted from the \href{https://github.com/lbenz730/ncaahoopR}{\code{ncaahoopR}} package by Luke Benz.
}
\usage{
dict
}
\description{
A lookup table of team name formats across major NCAA basketball data sources.
}
\details{
Useful for converting team names between different systems and data sources.
}
\examples{
data(dict)
head(dict)
}
\keyword{datasets}
