% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usu_data.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Team Name Conversion Dictionary (from ncaahoopR) with added column to
address some common naming errors}
\format{
A data frame with 10 columns:
\describe{
\item{NCAA}{Team name as listed on the NCAA website}
\item{ESPN}{Team name as listed in ESPN URLs}
\item{ESPN_PBP}{Team name as listed on ESPN play-by-play logs}
\item{Warren_Nolan}{Team name from WarrenNolan.com}
\item{Trank}{Team name from barttorvik.com}
\item{name_247}{Team name from 247Sports.com}
\item{conference}{The team's conference}
\item{sref_link}{Slug for the team on sports-reference.com}
\item{sref_name}{Team name on sports-reference.com}
\item{short_name}{Abbreviated team name (e.g., "Boise St"
instead of "Boise State")}
}
}
\source{
Extracted from
the \href{https://github.com/lbenz730/ncaahoopR}{\code{ncaahoopR}} package by
Luke Benz.
}
\usage{
dictionary
}
\description{
A lookup table of team name formats across major NCAA basketball data
sources.
}
\details{
Useful for converting team names between different systems and data sources.
}
\examples{
data(dictionary)
head(dictionary)
}
\keyword{datasets}
