% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refstats.R
\name{plot_fouls}
\alias{plot_fouls}
\title{Plot total fouls per game by referee}
\usage{
plot_fouls(foul_stats, top_n = 5)
}
\arguments{
\item{foul_stats}{A data frame generated from \code{refstats()}.}

\item{top_n}{Number of top referees to display in the plot.}
}
\value{
A ggplot object.
}
\description{
Helper function that visualizes the top referees by average fouls called per
game.
}
