% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refstats.R
\name{plot_team_foul_ratio}
\alias{plot_team_foul_ratio}
\title{Plot foul ratio by referee}
\usage{
plot_team_foul_ratio(ref_stats, top_n = 5)
}
\arguments{
\item{ref_stats}{A data frame generated from \code{refstats()}.}

\item{top_n}{Number of top referees to display by foul ratio.}
}
\value{
A ggplot object.
}
\description{
Helper function to visualize the ratio of fouls called on the team versus
the opponent.
}
