% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refstats.R
\name{plot_team_fouls}
\alias{plot_team_fouls}
\title{Plot team vs opponent foul counts}
\usage{
plot_team_fouls(ref_stats, top_n = 5)
}
\arguments{
\item{ref_stats}{A data frame generated from \code{refstats()}.}

\item{top_n}{Number of top referees to include based on total foul counts.}
}
\value{
A ggplot object.
}
\description{
Helper function to generate a bar chart comparing fouls called on the team
vs. opponent for each referee.
}
